fix publishing bug on receiver and cleanup user interface
Firstly, the receiver had a bug where publishing of received messages
was found to be unreliable if a ros subscriber exists before the ros
publisher is created. This is circumvented by creating the publisher
with latch=True.

The other issue was that the receiver script was drawing a weird
terminal ui which interfered with normal ros output when the script was
started via a launch file. That terminal ui feature has been completely
removed. Due to this, the feature to remap one originating hosts message
to / was also removed.