feat(sender): dynamically get monitoring host ip
so that we can get it from the `game_settings.yaml` via the
parameter_blackboard node. We do this, so that we can configure the
monitoring host ip with the game settings script used by the
`robot_compile.py` script.

As utilizing `get_parameters_from_other_node` from bitbots_utils for
this lead to some execution issues with spin/callbacks the sender was
adapted to explicitly use an executor and timer.